+++ src/apps/Shop/pages/BuyCheckout/CartProduct/index.tsx
@@ -2,10 +2,10 @@ import {useDispatch} from 'react-redux';
 
 import ActionLink from 'apps/Shop/components/ActionLink';
 import Thumbnail from 'apps/Shop/components/Thumbnail';
+import ProductListDetails from 'apps/Shop/components/ProductListDetails';
 import {unsetCartProduct} from 'apps/Shop/store/cartProducts';
 import {Product} from 'apps/Shop/types';
 import {AppDispatch, SFC, ToastType} from 'system/types';
-import {truncate} from 'system/utils/strings';
 import {displayToast} from 'system/utils/toast';
 import * as S from './Styles';
 
@@ -16,7 +16,7 @@ export interface CartProductProps {
 const CartProduct: SFC<CartProductProps> = ({product}) => {
   const dispatch = useDispatch<AppDispatch>();
 
-  const {description, imageUrl, name, productId} = product;
+  const {imageUrl, productId} = product;
 
   const handleRemoveClick = () => {
     dispatch(unsetCartProduct(productId));
@@ -26,11 +26,7 @@ const CartProduct: SFC<CartProductProps> = ({product}) => {
   return (
     <>
       <Thumbnail thumbnailUrl={imageUrl} />
-      <S.Details>
-        <S.Name>{name}</S.Name>
-        <S.Description>{truncate(description, 200)}</S.Description>
-        <S.PriceMini product={product} />
-      </S.Details>
+      <ProductListDetails product={product} />
       <S.Actions>
         <ActionLink onClick={handleRemoveClick}>Remove</ActionLink>
       </S.Actions>
